+++ R/processor.R
@@ -33,12 +33,11 @@
 #' @export
 exec = function(pipeline, on, with = NULL, ...)
 {
-  async_com = ""
-  json_file = ""
-
   # Parse options and give precedence to 1. global options 2. LAScatalog 3. with arguments 4. ... arguments
   with = parse_options(on, with, ...)
 
+  # String to a file. Special case for experimental lasRui: github.com/r-lidar/lasRui
+  # No documented, experimental, very hacky
   if (is.character(pipeline))
   {
     if (!file.exists(pipeline))
@@ -49,8 +48,9 @@ exec = function(pipeline, on, with = NULL, ...)
     processed_content <- do.call(c, processed_content)
     json_file <- tempfile(fileext = ".json")
     writeLines(processed_content, json_file)
-    async_com = tempfile(fileext = ".tmp")
+    with$progress_file = tempfile(fileext = ".tmp")
   }
+  # Normal case
   else if (methods::is(pipeline, "PipelinePtr"))
   {
     on_is_valid = FALSE
@@ -102,6 +102,7 @@ exec = function(pipeline, on, with = NULL, ...)
       on_is_valid = TRUE
     }
 
+    # Special case for the R API
     if (methods::is(on, "lasrcloud"))
     {
       xptr = on[[1]]
@@ -116,7 +117,7 @@ exec = function(pipeline, on, with = NULL, ...)
     json_file <- .APIOPERATIONS$generate_json(pipeline, on, with)
   }
 
-  ans  <- .APIOPERATIONS$excecute_pipeline(json_file, async_com)
+  ans  <- .APIOPERATIONS$excecute_pipeline(json_file)
 
   if (!with$noread) ans <- read_as_common_r_object(ans)
   ans <- Filter(Negate(is.null), ans)
@@ -188,40 +189,55 @@ parse_options = function(on, with, ...)
   noprocess <- NULL
   verbose <- FALSE
   noread <- FALSE
-  profiling <- ""
+  profile_file <- ""
+  progress_file <- ""
+  log_file <- ""
 
   # Explicit options
-  if (!is.null(dots$buffer)) buffer <- dots$buffer
-  if (!is.null(dots$chunk))  chunk <- dots$chunk
-  if (!is.null(dots$progress)) progress <- dots$progress
-  if (!is.null(dots$ncores)) ncores <- dots$ncores
-  if (!is.null(dots$noprocess)) noprocess <- dots$noprocess
-  if (!is.null(dots$verbose)) verbose <- dots$verbose
-  if (!is.null(dots$noread)) noread <- dots$noread
-  if (!is.null(dots$profiling)) profiling <- dots$profiling
+  if (!is.null(dots[["buffer"]])) buffer <- dots[["buffer"]]
+  if (!is.null(dots[["chunk"]]))  chunk <- dots[["chunk"]]
+  if (!is.null(dots[["progress"]])) progress <- dots[["progress"]]
+  if (!is.null(dots[["ncores"]])) ncores <- dots[["ncores"]]
+  if (!is.null(dots[["noprocess"]])) noprocess <- dots[["noprocess"]]
+  if (!is.null(dots[["verbose"]])) verbose <- dots[["verbose"]]
+  if (!is.null(dots[["noread"]])) noread <- dots[["noread"]]
+  if (!is.null(dots[["profile_file"]])) profile_file <- dots[["profile_file"]]
+  if (!is.null(dots[["progress_file"]])) progress_file <- dots[["progress_file"]]
+  if (!is.null(dots[["log_file"]])) log_file <- dots[["log_file"]]
 
   # 'with' list has precedence
-  if (!is.null(with$buffer)) buffer <- with$buffer
-  if (!is.null(with$chunk))  chunk <- with$chunk
-  if (!is.null(with$progress)) progress <- with$progress
-  if (!is.null(with$ncores))  ncores <- with$ncores
-  if (!is.null(with$noprocess)) noprocess <- with$noprocess
-  if (!is.null(with$verbose)) verbose <- with$verbose
-  if (!is.null(with$noread)) noread <- with$noread
-  if (!is.null(with$profiling)) profiling <- with$profiling
+  if (!is.null(with[["buffer"]])) buffer <- with[["buffer"]]
+  if (!is.null(with[["chunk"]]))  chunk <- with[["chunk"]]
+  if (!is.null(with[["progress"]])) progress <- with[["progress"]]
+  if (!is.null(with[["ncores"]]))  ncores <- with[["ncores"]]
+  if (!is.null(with[["noprocess"]])) noprocess <- with[["noprocess"]]
+  if (!is.null(with[["verbose"]])) verbose <- with[["verbose"]]
+  if (!is.null(with[["noread"]])) noread <- with[["noread"]]
+  if (!is.null(with[["profile_file"]])) profile_file <- with[["profile_file"]]
+  if (!is.null(with[["progress_file"]])) progress_file <- with[["progress_file"]]
+  if (!is.null(with[["log_file"]])) log_file <- with[["log_file"]]
 
   if (!missing(on))
   {
     # If 'on' is a LAScatalog it has the precedence on the 'with' list
     if (methods::is(on, "LAScatalog"))
     {
-      chunk <- on@chunk_options$size
-      buffer <- on@chunk_options$buffer
-      alignment <- on@chunk_options$alignment
-      progress <- on@processing_options$progress
-      processed <- on$processed
+      chunk <- on@chunk_options[["size"]]
+      buffer <- on@chunk_options[["buffer"]]
+      alignment <- on@chunk_options[["alignment"]]
+      progress <- on@processing_options[["progress"]]
+      processed <- on[["processed"]]
       if (!is.null(processed)) noprocess <- !processed
-      if (on@input_options$filter != "") warning(paste0("This LAScatalog has filter = \"", on@input_options$filter, "\" but this option is not automatically propagated to the 'reader_las()' stage.") , call. = FALSE, immediate. = TRUE)
+      if (on@input_options[["filter"]] != "")
+        warning(
+          paste0(
+            "This LAScatalog has filter = \"",
+            on@input_options[["filter"]],
+            "\" but this option is not automatically propagated to the 'reader_las()' stage."
+          ),
+          call. = FALSE,
+          immediate. = TRUE
+        )
     }
   }
 
@@ -233,13 +249,15 @@ parse_options = function(on, with, ...)
   mode <- match.arg(mode, modes)
 
   # Global options have precedence on everything
-  if (!is.null(LASROPTIONS$buffer)) buffer <- LASROPTIONS$buffer
-  if (!is.null(LASROPTIONS$chunk))  chunk <- LASROPTIONS$chunk
-  if (!is.null(LASROPTIONS$progress)) progress <- LASROPTIONS$progress
-  if (!is.null(LASROPTIONS$ncores)) ncores <- LASROPTIONS$ncores
-  if (!is.null(LASROPTIONS$strategy)) mode <- LASROPTIONS$strategy
-  if (!is.null(LASROPTIONS$verbose)) verbose <- LASROPTIONS$verbose
-  if (!is.null(LASROPTIONS$noread)) noread <- LASROPTIONS$noread
+  if (!is.null(LASROPTIONS[["buffer"]])) buffer <- LASROPTIONS[["buffer"]]
+  if (!is.null(LASROPTIONS[["chunk"]]))  chunk <- LASROPTIONS[["chunk"]]
+  if (!is.null(LASROPTIONS[["progress"]])) progress <- LASROPTIONS[["progress"]]
+  if (!is.null(LASROPTIONS[["ncores"]])) ncores <- LASROPTIONS[["ncores"]]
+  if (!is.null(LASROPTIONS[["strategy"]])) mode <- LASROPTIONS[["strategy"]]
+  if (!is.null(LASROPTIONS[["verbose"]])) verbose <- LASROPTIONS[["verbose"]]
+  if (!is.null(LASROPTIONS[["noread"]])) noread <- LASROPTIONS[["noread"]]
+  if (!is.null(LASROPTIONS[["progress_file"]])) progress_file <- LASROPTIONS[["progress_file"]]
+  if (!is.null(LASROPTIONS[["log_file"]])) log_file <- LASROPTIONS[["log_file"]]
 
   if (!has_omp_support())
   {
@@ -255,19 +273,27 @@ parse_options = function(on, with, ...)
   stopifnot(is.character(mode))
   stopifnot(is.logical(verbose))
   stopifnot(is.logical(noread))
+  stopifnot(is.character(progress_file))
+  stopifnot(is.character(log_file))
+  stopifnot(is.character(profile_file))
 
-  ret = list(ncores = ncores,
-             strategy = mode,
-             buffer = buffer,
-             progress = progress,
-             noprocess = noprocess,
-             chunk = chunk,
-             noread = noread,
-             verbose = verbose,
-             profiling = profiling)
+  ret = list(
+    ncores = ncores,
+    strategy = mode,
+    buffer = buffer,
+    progress = progress,
+    noprocess = noprocess,
+    chunk = chunk,
+    noread = noread,
+    verbose = verbose,
+    profile_file = profile_file,
+    progress_file = progress_file,
+    log_file = log_file
+  )
 
   return(ret)
 }
+
 
 #' Set global processing options
 #'