+++ NEWS.md
@@ -2,6 +2,15 @@
 
 - Fix: #198 LAS files with 0 points are discarded on-the-fly.
 - New: argument `store_in_attribute` in `local_maximum`
+- New: *undocumented* capacity at loggin information in files. Accessible in R API via
+  ```r
+  exec(..., progress_file = "path/to/progress.ext", log_file = "path/to/log.ext".
+  ```
+  Accessible in the C++ API via two members
+  ```r
+  Pipeline::set_progress_file(std::string);
+  Pipeline::set_profile_file(std::string);
+  ```
 
 # lasR 0.17.1